requirements.txt
Django>=4.0.6
python-dotenv>=0.20.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.3
requests>=2.28.1